odoo/requirements.txt
# Odoo 19 specific requirements
odoo>=19.0,<20.0

# Additional Python packages for Berit Loan module
Pillow>=10.0.0
lxml>=4.6.0
python-dateutil>=2.8.0
pytz>=2023.3
reportlab>=4.0.0
num2words>=0.5.12
ofxparse>=0.21
zeep>=4.2.0
qrcode>=7.4.0
vobject>=0.9.6
requests>=2.31.0
psycopg2-binary>=2.9.0
passlib>=1.7.4
babel>=2.12.0
html2text>=2020.1.16
matplotlib>=3.7.0
gevent>=23.7.0
